// frontend/src/api/generated/models/UpdateTodoItem.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Todo Microservice API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 *
 * @export
 * @interface UpdateTodoItem
 */
export interface UpdateTodoItem {
  /**
   *
   * @type {string}
   * @memberof UpdateTodoItem
   */
  title: string;
  /**
   *
   * @type {string}
   * @memberof UpdateTodoItem
   */
  description: string;
  /**
   *
   * @type {boolean}
   * @memberof UpdateTodoItem
   */
  completed: boolean;
  /**
   *
   * @type {Date}
   * @memberof UpdateTodoItem
   */
  dueDate?: Date;
  /**
   *
   * @type {string}
   * @memberof UpdateTodoItem
   */
  position: string;
}

/**
 * Check if a given object implements the UpdateTodoItem interface.
 */
export function instanceOfUpdateTodoItem(value: object): value is UpdateTodoItem {
  if (!('title' in value) || value['title'] === undefined) return false;
  if (!('description' in value) || value['description'] === undefined) return false;
  if (!('completed' in value) || value['completed'] === undefined) return false;
  if (!('position' in value) || value['position'] === undefined) return false;
  return true;
}

export function UpdateTodoItemFromJSON(json: any): UpdateTodoItem {
  return UpdateTodoItemFromJSONTyped(json, false);
}

export function UpdateTodoItemFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean
): UpdateTodoItem {
  if (json == null) {
    return json;
  }
  return {
    title: json['title'],
    description: json['description'],
    completed: json['completed'],
    dueDate: json['due_date'] == null ? undefined : new Date(json['due_date']),
    position: json['position'],
  };
}

export function UpdateTodoItemToJSON(json: any): UpdateTodoItem {
  return UpdateTodoItemToJSONTyped(json, false);
}

export function UpdateTodoItemToJSONTyped(
  value?: UpdateTodoItem | null,
  ignoreDiscriminator: boolean = false
): any {
  if (value == null) {
    return value;
  }

  return {
    title: value['title'],
    description: value['description'],
    completed: value['completed'],
    due_date: value['dueDate'] == null ? undefined : value['dueDate'].toISOString(),
    position: value['position'],
  };
}